+++ Flow/src/utopia/flow/caching/multi/TryCache.scala
@@ -1,6 +1,6 @@
 package utopia.flow.caching.multi
 
-import utopia.flow.caching.single.{ExpiringSingleCache, SingleTryCache}
+import utopia.flow.caching.single.SingleTryCache
 
 import scala.concurrent.duration.FiniteDuration
 import scala.util.Try
@@ -35,4 +35,17 @@ object TryCache
 	def expiring[Key, Value](failResultDuration: FiniteDuration, maxResultDuration: FiniteDuration)(
 		request: Key => Try[Value]) = ExpiringCache[Key, Try[Value]] {
 		key: Key => SingleTryCache.expiring(failResultDuration, maxResultDuration){ request(key) } }
+	
+	/**
+	  * Creates a new cache that releases its contents after a while, after which it still uses weak references
+	  * @param failureReleaseAfterDuration Duration to keep failure states strongly referenced
+	  * @param successReleaseAfterDuration Duration to keep success states strongly referenced
+	  * @param request A function for retrieving a result
+	  * @tparam Key Type of key for this cache
+	  * @tparam Value Success result type
+	  * @return A new cache
+	  */
+	def releasing[Key, Value](failureReleaseAfterDuration: FiniteDuration, successReleaseAfterDuration: FiniteDuration)
+							 (request: Key => Try[Value]) =
+		ReleasingCache(expiring(failureReleaseAfterDuration, successReleaseAfterDuration)(request))
 }