+++ Reflection/src/utopia/reflection/component/reach/template/ReachComponent.scala
@@ -27,9 +27,16 @@ trait ReachComponent extends ReachComponentLike with CachingStackable2
 	
 	// Whenever component bounds update, repaints the affected area
 	boundsPointer.addListener { event =>
-		// TODO: Use copy area instead of full repaint when size stays the same
-		Bounds.aroundOption(Vector(event.oldValue, event.newValue).filter { _.size.isPositive })
-			.foreach { parentHierarchy.repaint(_) }
+		// Case: Component size changed => repaints affected area (including both old and new bounds)
+		// TODO: Fix copy issue
+		// TODO: Also, leaves some of the old area unpainted (like one pixel to the right)
+		/*
+		if (event.compareBy { _.size })*/
+			Bounds.aroundOption(Vector(event.oldValue, event.newValue).filter { _.size.isPositive }/*.map { _.ceil }*/)
+				.foreach { parentHierarchy.repaint(_) }
+		// Case: Only position changed => shifts this component, copying already painted region
+		/*else if (event.newValue.size.isPositive)
+			paintMovement((event.newValue.position - event.oldValue.position).toVector)*/
 	}